Testing safe driver interaction with callbacks using extern
For the record, I'm so sorry. This is messy boilerplate code to test if it actually works and how much code we need to call a simple function. It works, and the unsurprising result when it comes to boilerplate is... well, it takes more than two lines...

So now that it works, I will create a macro and some util, to avoid writing all this boilerplate manually. It's still not going to be the most pretty thing, but that's sadly the best we can do right now.