Adding literal expressions
This splits the literal expression in to separate structs to deal with the specific requirements the separate literals have. I'm unsure if this is ideal, but I have been stuck on solving it another way for too long. We can still change this later if the need arises.

For me, splitting them, actually makes sense. I can't think of a (linting related) use case, that would want to handle string literals the same as integer literals. Combining numerical literals could make sense, but this makes the methods more expressive and the documentation more detailed. Some util methods might get a bit more verbose, but that's a price I'm willing to pay (famous last words)