Rustc driver: change alloc function signature
Previously, rustc's driver was using closures to call the `alloc` function of bumpalo as an optimization. However, this sadly convoluted the backtrace. The documentation of bumpalo also states that the usage of closures helps in some special edge cases, but using `alloc()` should be equivalent in most cases. Therefore, I decided to remove the closures in favor of readability and easy of writing. Here is an example of the backtrace points, which have now been removed.

```
[...]
   6: bumpalo::Bump::alloc_with::inner_writer
   7: bumpalo::Bump::alloc_with
[...]
```