API: Add semantic type representation
Some more context: Originally, I hoped to combine the semantic and syntactic type representation into one. However, after looking at the internal representation of the two and some more consideration, I came to the conclusion that this would be a bad idea. Here are a few reasons for this change:
* Util functions could not distinguish semantic and syntactic types, based on the data type. Instead, we would need to use runtime checks.
* The representations are too different. A syntactic path type can be several different semantic things. This adds complicated edge cases.
* The documentation can now be adjusted, to include semantic or syntactic specific information

This commit also introduces a new node representation. While I still love the idea of mainly using `Kind` enums, as proposed in `rust-marker/design#13` I'm starting to see flaws. In favor of stability and using the rigid type system, I believe it's better to use a parent node. In this case, `SemTy` is added along with `SemTyKind`.