Refactoring(API): Remove `lint_pass_fns` macro
Using a macro to generate this part of the API, was a good idea, but it sadly wasn't flexible enough. Additionally, it had some problems with auto-completion in lint crates. Writing all of the boilerplate manually gives more flexibility for now. In the future, I would like to generate a good part of this code based on the `LintPass` trait definition, but that's a project for a rainy day.