Fix quemu setup using to new curl options
The build failed in driver ARM build due to the usage of `--retry-all-errors`.

The build on master also failed, because we don't use this option on `ubuntu20.04`. I think we'll need to add a manual retry loop there as a crutch for old systems with old curl.

As for now I'll move the `v0`, `v0.4` and `v0.4.0` to this new commit and this will trigger the "release-on-tag" workflow again