Fix marker_rustc_driver build ARM just one more time
When last time I fixed the build of marker_rustc_driver on ARM I tested my changes only locally in an isolated folder where I manually "unit-tested" the `upload.sh` script. Of course, when the build happens on CI the folders layout is entirely different. That's why the ARM packaging script broke but in other way.

On CI we mount the `.` fodler as `/artifacts` in `quemu`, but `.` folder includes the `marker_rustc_driver` crate source code folder. So when we do `cp target/release/marker_rustc_driver /artifacts/marker_rustc_driver` we copy the binary inside of that folder...

I fixed this by mounting `./artifacts` which is guaranteed to be an empty folder