Remove params that aren't supported by `cargo release publish`
I was confused by [this warning on CI](https://github.com/rust-marker/marker/actions/runs/6984551469/job/19007537021#step:6:14)

It looked to me as if cargo release wanted to do a commit and push changes to the remote. I still don't understand why that warning is emitted because `cargo relase` is used only for publishing. Maybe that's a micro-bug in `cargo-release`. Anyway, this time it was me who broke crates publishing. Hopefully, this was the last obstacle